02-code-review-agent/requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
openai>=1.50.0
python-dotenv>=1.0.0
pygments>=2.17.0
gitpython>=3.1.40
